requirements.txt
RPi.BME280==0.2.4
smbus2==0.4.2
Adafruit-Blinka
adafruit_ht16k33

